+++ CHANGELOG.md
@@ -1,9 +1,21 @@
 ## 2.0.0 (Work in progress)
 
-- Bumped the minimum SDK version to `2.17.0`.
-- Removed dependencies:
-  - `sprint`
-  - `web_scraper`
+- Additions:
+  - Added the `meta` package for static analysis.
+  - Added support for the 'Sitemap' field.
+  - Added support for specifying:
+    - The precedent rule type for determining whether a certain user-agent can
+      or cannot access a certain path. (`PrecedentRuleType`)
+    - The comparison strategy to use for comparing rule precedence.
+      (`PrecedenceStrategy`)
+- Changes:
+  - Bumped the minimum SDK version to `2.17.0`.
+- Improvements:
+  - Made all structs `const` and marked them as `@sealed` and `@immutable`.
+- Deletions:
+  - Removed dependencies:
+    - `sprint`
+    - `web_scraper`
 
 ## 1.1.1